+++ src/applications/disability-benefits/526EZ/pages/primaryAddress.js
@@ -22,6 +22,8 @@ import  {
   STATE_LABELS,
   STATE_VALUES,
   USA,
+} from '../../all-claims/constants';
+import  {
   ADDRESS_TYPES
 } from '../constants';
